Factor out UriField and SchemaIndexFields from DocTypeBuilder.
