+++ searchlib/src/vespa/searchlib/index/doctypebuilder.cpp
@@ -68,158 +68,6 @@ insertStructType(document::DocumenttypesConfig::Documenttype & cfg,
 
 }
 
-DocTypeBuilder::UriField::UriField()
-    : _all(Schema::UNKNOWN_FIELD_ID),
-      _scheme(Schema::UNKNOWN_FIELD_ID),
-      _host(Schema::UNKNOWN_FIELD_ID),
-      _port(Schema::UNKNOWN_FIELD_ID),
-      _path(Schema::UNKNOWN_FIELD_ID),
-      _query(Schema::UNKNOWN_FIELD_ID),
-      _fragment(Schema::UNKNOWN_FIELD_ID),
-      _hostname(Schema::UNKNOWN_FIELD_ID)
-{
-}
-
-
-bool
-DocTypeBuilder::UriField::valid(const Schema &schema,
-                                uint32_t fieldId,
-                                const Schema::CollectionType &collectionType)
-{
-    if (fieldId == Schema::UNKNOWN_FIELD_ID)
-        return false;
-    const Schema::IndexField &field = schema.getIndexField(fieldId);
-    if (field.getDataType() != schema::DataType::STRING)
-        return false;
-    if (field.getCollectionType() != collectionType)
-        return false;
-    return true;
-}
-
-
-bool
-DocTypeBuilder::UriField::broken(const Schema &schema,
-                                 const Schema::CollectionType &
-                                 collectionType) const
-{
-    return !valid(schema, _all, collectionType) &&
-        valid(schema, _scheme, collectionType) &&
-        valid(schema, _host, collectionType) &&
-        valid(schema, _port, collectionType) &&
-        valid(schema, _path, collectionType) &&
-        valid(schema, _query, collectionType) &&
-        valid(schema, _fragment, collectionType);
-}
-
-bool
-DocTypeBuilder::UriField::valid(const Schema &schema,
-                                const Schema::CollectionType &
-                                collectionType) const
-{
-    return valid(schema, _all, collectionType) &&
-        valid(schema, _scheme, collectionType) &&
-        valid(schema, _host, collectionType) &&
-        valid(schema, _port, collectionType) &&
-        valid(schema, _path, collectionType) &&
-        valid(schema, _query, collectionType) &&
-        valid(schema, _fragment, collectionType);
-}
-
-
-void
-DocTypeBuilder::UriField::setup(const Schema &schema,
-                                const vespalib::string &field)
-{
-    _all = schema.getIndexFieldId(field);
-    _scheme = schema.getIndexFieldId(field + ".scheme");
-    _host = schema.getIndexFieldId(field + ".host");
-    _port = schema.getIndexFieldId(field + ".port");
-    _path = schema.getIndexFieldId(field + ".path");
-    _query = schema.getIndexFieldId(field + ".query");
-    _fragment = schema.getIndexFieldId(field + ".fragment");
-    _hostname = schema.getIndexFieldId(field + ".hostname");
-}
-
-
-void
-DocTypeBuilder::UriField::markUsed(UsedFieldsMap &usedFields,
-                                   uint32_t field)
-{
-    if (field == Schema::UNKNOWN_FIELD_ID)
-        return;
-    assert(usedFields.size() > field);
-    usedFields[field] = true;
-}
-
-
-void
-DocTypeBuilder::UriField::markUsed(UsedFieldsMap &usedFields) const
-{
-    markUsed(usedFields, _all);
-    markUsed(usedFields, _scheme);
-    markUsed(usedFields, _host);
-    markUsed(usedFields, _port);
-    markUsed(usedFields, _path);
-    markUsed(usedFields, _query);
-    markUsed(usedFields, _fragment);
-    markUsed(usedFields, _hostname);
-}
-
-
-
-DocTypeBuilder::SchemaIndexFields::SchemaIndexFields()
-    : _textFields(),
-      _uriFields()
-{
-}
-
-DocTypeBuilder::SchemaIndexFields::~SchemaIndexFields() {}
-
-void
-DocTypeBuilder::SchemaIndexFields::setup(const Schema &schema)
-{
-    uint32_t numIndexFields = schema.getNumIndexFields();
-    UsedFieldsMap usedFields;
-    usedFields.resize(numIndexFields);
-
-    // Detect all URI fields (flattened structs).
-    for (uint32_t fieldId = 0; fieldId < numIndexFields; ++fieldId) {
-        const Schema::IndexField &field = schema.getIndexField(fieldId);
-        const vespalib::string &name = field.getName();
-        size_t dotPos = name.find('.');
-        if (dotPos != vespalib::string::npos) {
-            const vespalib::string suffix = name.substr(dotPos + 1);
-            if (suffix == "scheme") {
-                const vespalib::string shortName = name.substr(0, dotPos);
-                UriField uriField;
-                uriField.setup(schema, shortName);
-                if (uriField.valid(schema, field.getCollectionType())) {
-                    _uriFields.push_back(uriField);
-                    uriField.markUsed(usedFields);
-                } else if (uriField.broken(schema,
-                                   field.getCollectionType())) {
-                    // Broken removal of unused URI fields.
-                    uriField.markUsed(usedFields);
-                }
-            }
-        }
-    }
-
-    // Non-URI fields are currently supposed to be text fields.
-    for (uint32_t fieldId = 0; fieldId < numIndexFields; ++fieldId) {
-        if (usedFields[fieldId])
-            continue;
-        const Schema::IndexField &field = schema.getIndexField(fieldId);
-        switch (field.getDataType()) {
-        case schema::DataType::STRING:
-            _textFields.push_back(fieldId);
-            break;
-        default:
-            ;
-        }
-    }
-}
-
 DocTypeBuilder::DocTypeBuilder(const Schema &schema)
     : _schema(schema),
       _iFields()
